requirements.txt
flask==3.1.1
sqlalchemy==2.0.41
mysql-connector-python==9.2.0
azure-identity==1.23.0
azure-keyvault-secrets==4.10.0